Refactor: Gamewebsocket.js 파일 이동
