+++ FE/src/pages/GameSettingPage.js
@@ -1,7 +1,7 @@
 import { changeUrlData } from '../index.js';
 import HorizontalButton from '../components/HorizontalButton.js';
 import VerticalButton from '../components/VerticalButton.js';
-import { Gamewebsocket } from '../Gamewebsocket.js';
+import { Gamewebsocket } from '../websocket/Gamewebsocket.js';
 
 const html = String.raw;
